+++ trilha_desafio_5/Livraria/api/src/controllers/editoraController.js
@@ -1,11 +1,12 @@
-import editoraService from "../services/editoraService.js";
+import editoraService from "../models/services/editoraService.js";
+import isValidDados from "../models/validations/validaDados.js";
 
 const getAllEditora = async (req, res) => {
   try {
     const editoras = await editoraService.getAllEditora();
 
     if (editoras.length === 0) {
-      return res.status(404).json({ message: "Nenhum editora encontrado." });
+      return res.status(404).json({ message: "Nenhuma editora encontrada." });
     }
 
     return res.status(200).json(editoras);
@@ -22,7 +23,7 @@ const getIdEditora = async (req, res) => {
     const editora = await editoraService.getIdEditora(req.params.id);
 
     if (!editora) {
-      return res.status(404).json({ message: "Editora não encontrado." });
+      return res.status(404).json({ message: "Editora não encontrada." });
     }
     return res.status(200).json(editora);
   } catch (error) {
@@ -33,18 +34,79 @@ const getIdEditora = async (req, res) => {
 };
 
 const postEditora = async (req, res) => {
-  await editoraService.postEditora(req.body);
-  return res.sendStatus(201);
+  try {
+    if (Object.keys(req.body).length === 0) {
+      return res
+        .status(400)
+        .json({ message: "Os dados da editora não foram fornecidos." });
+    }
+
+    const reqBodyNome = req.body.nome.trim();
+    req.body.localizacao = req.body.localizacao.trim();
+
+    if (!isValidDados.isValidDadosEditora(req.body)) {
+      return res
+        .status(400)
+        .json({ message: "Os dados da editora são inválidos." });
+    }
+
+    const existeEditora = await editoraService.getEditoraByName(reqBodyNome);
+    if (existeEditora) {
+      return res
+        .status(400)
+        .json({ message: "Esta editora já está cadastrada." });
+    }
+
+    await editoraService.postEditora(req.body);
+
+    return res.sendStatus(201);
+  } catch (error) {
+    console.error("Erro ao tentar adicionar editora:", error);
+    return res.status(500).json({ error: "Erro interno do servidor" });
+  }
 };
 
 const patchEditora = async (req, res) => {
-  await editoraService.patchEditora(req.params.id, req.body);
-  return res.sendStatus(200);
+  try {
+    if (Object.keys(req.body).length === 0) {
+      return res
+        .status(400)
+        .json({ message: "Os dados da editora não foram fornecidos." });
+    }
+
+    if (!isValidDados.isValidDadosEditora(req.body)) {
+      return res
+        .status(400)
+        .json({ message: "Os novos dados da editora são inválidos." });
+    }
+
+    const existeEditora = await editoraService.getIdEditora(req.params.id);
+    if (!existeEditora) {
+      return res.status(404).json({ message: "Editora não encontrada." });
+    }
+
+    await editoraService.patchEditora(req.params.id, req.body);
+
+    return res.sendStatus(200);
+  } catch (error) {
+    console.error("Erro ao tentar atualizar editora:", error);
+    return res.status(500).json({ error: "Erro interno do servidor" });
+  }
 };
 
 const deleteEditora = async (req, res) => {
-  await editoraService.deleteEditora(req.params.id);
-  return res.sendStatus(204);
+  try {
+    const existeEditora = await editoraService.getIdEditora(req.params.id);
+    if (!existeEditora) {
+      return res.status(404).json({ message: "Editora não encontrada." });
+    }
+
+    await editoraService.deleteEditora(req.params.id);
+    return res.status(200).json({ message: "Editora deletada com sucesso." });
+  } catch (error) {
+    console.error("Erro ao tentar excluir editora:", error);
+    return res.status(500).json({ error: "Erro interno do servidor" });
+  }
 };
 
 export default {
